Refactor search action to extract shared functions
